+++ pyg_lib/csrc/sampler/cpu/neighbor_kernel.cpp
@@ -72,14 +72,14 @@ class NeighborSampler {
             dst_mapper, generator, out_global_dst_nodes);
   }
 
-  void temporal_sample(const node_t global_src_node,
-                       const scalar_t local_src_node,
-                       const int64_t count,
-                       const temporal_t seed_time,
-                       const temporal_t* time,
-                       pyg::sampler::Mapper<node_t, scalar_t>& dst_mapper,
-                       pyg::random::RandintEngine<scalar_t>& generator,
-                       std::vector<node_t>& out_global_dst_nodes) {
+  void node_temporal_sample(const node_t global_src_node,
+                            const scalar_t local_src_node,
+                            const int64_t count,
+                            const temporal_t seed_time,
+                            const temporal_t* time,
+                            pyg::sampler::Mapper<node_t, scalar_t>& dst_mapper,
+                            pyg::random::RandintEngine<scalar_t>& generator,
+                            std::vector<node_t>& out_global_dst_nodes) {
     auto row_start = rowptr_[to_scalar_t(global_src_node)];
     auto row_end = rowptr_[to_scalar_t(global_src_node) + 1];
 
@@ -98,6 +98,38 @@ class NeighborSampler {
 
     if (row_end - row_start > 1) {
       TORCH_CHECK(time[col_[row_start]] <= time[col_[row_end - 1]],
+                  "Found invalid non-sorted temporal neighborhood");
+    }
+
+    _sample(global_src_node, local_src_node, row_start, row_end, count,
+            dst_mapper, generator, out_global_dst_nodes);
+  }
+
+  void edge_temporal_sample(const node_t global_src_node,
+                            const scalar_t local_src_node,
+                            const int64_t count,
+                            const temporal_t seed_time,
+                            const temporal_t* time,
+                            pyg::sampler::Mapper<node_t, scalar_t>& dst_mapper,
+                            pyg::random::RandintEngine<scalar_t>& generator,
+                            std::vector<node_t>& out_global_dst_nodes) {
+    auto row_start = rowptr_[to_scalar_t(global_src_node)];
+    auto row_end = rowptr_[to_scalar_t(global_src_node) + 1];
+
+    if ((row_end - row_start == 0) || (count == 0))
+      return;
+
+    // Find new `row_end` such that all neighbors fulfill temporal constraints:
+    auto it = std::upper_bound(
+        time + row_start, time + row_end, seed_time,
+        [&](const scalar_t& a, const scalar_t& b) { return a < b; });
+    row_end = it - time;
+
+    if (temporal_strategy_ == "last" && count >= 0) {
+      row_start = std::max(row_start, (scalar_t)(row_end - count));
+    }
+    if (row_end - row_start > 1) {
+      TORCH_CHECK(time[row_start] <= time[row_end - 1],
                   "Found invalid non-sorted temporal neighborhood");
     }
 
@@ -307,26 +339,39 @@ sample(const at::Tensor& rowptr,
        const at::Tensor& col,
        const at::Tensor& seed,
        const std::vector<int64_t>& num_neighbors,
-       const c10::optional<at::Tensor>& time,
+       const c10::optional<at::Tensor>& node_time,
+       const c10::optional<at::Tensor>& edge_time,
        const c10::optional<at::Tensor>& seed_time,
        const c10::optional<at::Tensor>& edge_weight,
        const bool csc,
        const std::string temporal_strategy) {
-  TORCH_CHECK(!time.has_value() || disjoint,
+  TORCH_CHECK(!node_time.has_value() || disjoint,
               "Temporal sampling needs to create disjoint subgraphs");
+  TORCH_CHECK(!edge_time.has_value() || disjoint,
+              "Temporal sampling needs to create disjoint subgraphs");
+  TORCH_CHECK(!(node_time.has_value() && edge_time.has_value()),
+              "Only one of node-level or edge-level sampling is supported ");
 
   TORCH_CHECK(rowptr.is_contiguous(), "Non-contiguous 'rowptr'");
   TORCH_CHECK(col.is_contiguous(), "Non-contiguous 'col'");
   TORCH_CHECK(seed.is_contiguous(), "Non-contiguous 'seed'");
-  if (time.has_value()) {
-    TORCH_CHECK(time.value().is_contiguous(), "Non-contiguous 'time'");
+  if (node_time.has_value()) {
+    TORCH_CHECK(node_time.value().is_contiguous(),
+                "Non-contiguous 'node_time'");
+  }
+  if (edge_time.has_value()) {
+    TORCH_CHECK(edge_time.value().is_contiguous(),
+                "Non-contiguous 'edge_time'");
+    TORCH_CHECK(seed_time.has_value(), "Seed time needs to be specified");
   }
   if (seed_time.has_value()) {
     TORCH_CHECK(seed_time.value().is_contiguous(),
                 "Non-contiguous 'seed_time'");
   }
-  TORCH_CHECK(!(time.has_value() && edge_weight.has_value()),
-              "Biased temporal sampling not yet supported");
+  TORCH_CHECK(!(node_time.has_value() && edge_weight.has_value()),
+              "Biased node temporal sampling not yet supported");
+  TORCH_CHECK(!(edge_time.has_value() && edge_weight.has_value()),
+              "Biased edge temporal sampling not yet supported");
 
   at::Tensor out_row, out_col, out_node_id;
   c10::optional<at::Tensor> out_edge_id = c10::nullopt;
@@ -368,8 +413,8 @@ sample(const at::Tensor& rowptr,
         for (size_t i = 0; i < seed.numel(); ++i) {
           seed_times.push_back(seed_time_data[i]);
         }
-      } else if (time.has_value()) {
-        const auto time_data = time.value().data_ptr<temporal_t>();
+      } else if (node_time.has_value()) {
+        const auto time_data = node_time.value().data_ptr<temporal_t>();
         for (size_t i = 0; i < seed.numel(); ++i) {
           seed_times.push_back(time_data[seed_data[i]]);
         }
@@ -395,7 +440,7 @@ sample(const at::Tensor& rowptr,
           if constexpr (distributed)
             cumsum_neighbors_per_node.push_back(sampled_nodes.size());
         }
-      } else if (!time.has_value()) {
+      } else if (!node_time.has_value() && !edge_time.has_value()) {
         for (size_t i = begin; i < end; ++i) {
           sampler.uniform_sample(
               /*global_src_node=*/sampled_nodes[i],
@@ -408,19 +453,38 @@ sample(const at::Tensor& rowptr,
             cumsum_neighbors_per_node.push_back(sampled_nodes.size());
         }
       } else if constexpr (!std::is_scalar<node_t>::value) {  // Temporal:
-        const auto time_data = time.value().data_ptr<temporal_t>();
-        for (size_t i = begin; i < end; ++i) {
-          const auto batch_idx = sampled_nodes[i].first;
-          sampler.temporal_sample(
-              /*global_src_node=*/sampled_nodes[i],
-              /*local_src_node=*/i, /*count=*/count,
-              /*seed_time=*/seed_times[batch_idx],
-              /*time=*/time_data,
-              /*dst_mapper=*/mapper,
-              /*generator=*/generator,
-              /*out_global_dst_nodes=*/sampled_nodes);
-          if constexpr (distributed)
-            cumsum_neighbors_per_node.push_back(sampled_nodes.size());
+        if (edge_time.has_value()) {
+          const auto edge_time_data = edge_time.value().data_ptr<temporal_t>();
+          for (size_t i = begin; i < end; ++i) {
+            const auto batch_idx = sampled_nodes[i].first;
+            sampler.edge_temporal_sample(
+                /*global_src_node=*/sampled_nodes[i],
+                /*local_src_node=*/i,
+                /*count=*/count,
+                /*seed_time=*/seed_times[batch_idx],
+                /*time=*/edge_time_data,
+                /*dst_mapper=*/mapper,
+                /*generator=*/generator,
+                /*out_global_dst_nodes=*/sampled_nodes);
+            if constexpr (distributed)
+              cumsum_neighbors_per_node.push_back(sampled_nodes.size());
+          }
+        } else {
+          const auto node_time_data = node_time.value().data_ptr<temporal_t>();
+          for (size_t i = begin; i < end; ++i) {
+            const auto batch_idx = sampled_nodes[i].first;
+            sampler.node_temporal_sample(
+                /*global_src_node=*/sampled_nodes[i],
+                /*local_src_node=*/i,
+                /*count=*/count,
+                /*seed_time=*/seed_times[batch_idx],
+                /*time=*/node_time_data,
+                /*dst_mapper=*/mapper,
+                /*generator=*/generator,
+                /*out_global_dst_nodes=*/sampled_nodes);
+            if constexpr (distributed)
+              cumsum_neighbors_per_node.push_back(sampled_nodes.size());
+          }
         }
       }
       begin = end, end = sampled_nodes.size();
@@ -462,13 +526,18 @@ sample(const std::vector<node_type>& node_types,
        const c10::Dict<rel_type, at::Tensor>& col_dict,
        const c10::Dict<node_type, at::Tensor>& seed_dict,
        const c10::Dict<rel_type, std::vector<int64_t>>& num_neighbors_dict,
-       const c10::optional<c10::Dict<node_type, at::Tensor>>& time_dict,
+       const c10::optional<c10::Dict<node_type, at::Tensor>>& node_time_dict,
+       const c10::optional<c10::Dict<rel_type, at::Tensor>>& edge_time_dict,
        const c10::optional<c10::Dict<node_type, at::Tensor>>& seed_time_dict,
        const c10::optional<c10::Dict<rel_type, at::Tensor>>& edge_weight_dict,
        const bool csc,
        const std::string temporal_strategy) {
-  TORCH_CHECK(!time_dict.has_value() || disjoint,
-              "Temporal sampling needs to create disjoint subgraphs");
+  TORCH_CHECK(!node_time_dict.has_value() || disjoint,
+              "Node temporal sampling needs to create disjoint subgraphs");
+  TORCH_CHECK(!edge_time_dict.has_value() || disjoint,
+              "Edge temporal sampling needs to create disjoint subgraphs");
+  TORCH_CHECK(!(node_time_dict.has_value() && edge_time_dict.has_value()),
+              "Only one of node-level or edge-level sampling is supported ");
 
   for (const auto& kv : rowptr_dict) {
     const at::Tensor& rowptr = kv.value();
@@ -482,11 +551,18 @@ sample(const std::vector<node_type>& node_types,
     const at::Tensor& seed = kv.value();
     TORCH_CHECK(seed.is_contiguous(), "Non-contiguous 'seed'");
   }
-  if (time_dict.has_value()) {
-    for (const auto& kv : time_dict.value()) {
+  if (node_time_dict.has_value()) {
+    for (const auto& kv : node_time_dict.value()) {
       const at::Tensor& time = kv.value();
-      TORCH_CHECK(time.is_contiguous(), "Non-contiguous 'time'");
+      TORCH_CHECK(time.is_contiguous(), "Non-contiguous 'node_time'");
     }
+  }
+  if (edge_time_dict.has_value()) {
+    for (const auto& kv : edge_time_dict.value()) {
+      const at::Tensor& time = kv.value();
+      TORCH_CHECK(time.is_contiguous(), "Non-contiguous 'edge_time'");
+    }
+    TORCH_CHECK(seed_time_dict.has_value(), "Seed time needs to be specified");
   }
   if (seed_time_dict.has_value()) {
     for (const auto& kv : seed_time_dict.value()) {
@@ -494,7 +570,9 @@ sample(const std::vector<node_type>& node_types,
       TORCH_CHECK(seed_time.is_contiguous(), "Non-contiguous 'seed_time'");
     }
   }
-  TORCH_CHECK(!(time_dict.has_value() && edge_weight_dict.has_value()),
+  TORCH_CHECK(!(node_time_dict.has_value() && edge_weight_dict.has_value()),
+              "Biased temporal sampling not yet supported");
+  TORCH_CHECK(!(edge_time_dict.has_value() && edge_weight_dict.has_value()),
               "Biased temporal sampling not yet supported");
 
   c10::Dict<rel_type, at::Tensor> out_row_dict, out_col_dict;
@@ -604,8 +682,8 @@ sample(const std::vector<node_type>& node_types,
           for (size_t i = 0; i < seed.numel(); ++i) {
             seed_times.push_back(seed_time_data[i]);
           }
-        } else if (time_dict.has_value()) {
-          const at::Tensor& time = time_dict.value().at(kv.key());
+        } else if (node_time_dict.has_value()) {
+          const at::Tensor& time = node_time_dict.value().at(kv.key());
           const auto time_data = time.data_ptr<scalar_t>();
           seed_times.reserve(seed_times.size() + seed.numel());
           for (size_t i = 0; i < seed.numel(); ++i) {
@@ -659,8 +737,10 @@ sample(const std::vector<node_type>& node_types,
                         /*generator=*/generator,
                         /*out_global_dst_nodes=*/dst_sampled_nodes);
                   }
-                } else if (!time_dict.has_value() ||
-                           !time_dict.value().contains(dst)) {
+                } else if ((!node_time_dict.has_value() ||
+                            !node_time_dict.value().contains(dst)) &&
+                           (!edge_time_dict.has_value() ||
+                            !edge_time_dict.value().contains(to_rel_type(k)))) {
                   for (size_t i = begin; i < end; ++i) {
                     sampler.uniform_sample(
                         /*global_src_node=*/src_sampled_nodes[i],
@@ -671,20 +751,41 @@ sample(const std::vector<node_type>& node_types,
                         /*out_global_dst_nodes=*/dst_sampled_nodes);
                   }
                 } else if constexpr (!std::is_scalar<node_t>::value) {
-                  // Temporal sampling:
-                  const at::Tensor& dst_time = time_dict.value().at(dst);
-                  const auto dst_time_data = dst_time.data_ptr<temporal_t>();
-                  for (size_t i = begin; i < end; ++i) {
-                    const auto batch_idx = src_sampled_nodes[i].first;
-                    sampler.temporal_sample(
-                        /*global_src_node=*/src_sampled_nodes[i],
-                        /*local_src_node=*/i,
-                        /*count=*/count,
-                        /*seed_time=*/seed_times[batch_idx],
-                        /*time=*/dst_time_data,
-                        /*dst_mapper=*/dst_mapper,
-                        /*generator=*/generator,
-                        /*out_global_dst_nodes=*/dst_sampled_nodes);
+                  if (edge_time_dict.has_value() &&
+                      edge_time_dict.value().contains(to_rel_type(k))) {
+                    // Edge-level temporal sampling:
+                    const at::Tensor& edge_time =
+                        edge_time_dict.value().at(to_rel_type(k));
+                    const auto edge_time_data =
+                        edge_time.data_ptr<temporal_t>();
+                    for (size_t i = begin; i < end; ++i) {
+                      const auto batch_idx = src_sampled_nodes[i].first;
+                      sampler.edge_temporal_sample(
+                          /*global_src_node=*/src_sampled_nodes[i],
+                          /*local_src_node=*/i,
+                          /*count=*/count,
+                          /*seed_time=*/seed_times[batch_idx],
+                          /*time=*/edge_time_data,
+                          /*dst_mapper=*/dst_mapper,
+                          /*generator=*/generator,
+                          /*out_global_dst_nodes=*/dst_sampled_nodes);
+                    }
+                  } else {
+                    // Node-level temporal sampling:
+                    const at::Tensor& dst_time = node_time_dict.value().at(dst);
+                    const auto dst_time_data = dst_time.data_ptr<temporal_t>();
+                    for (size_t i = begin; i < end; ++i) {
+                      const auto batch_idx = src_sampled_nodes[i].first;
+                      sampler.node_temporal_sample(
+                          /*global_src_node=*/src_sampled_nodes[i],
+                          /*local_src_node=*/i,
+                          /*count=*/count,
+                          /*seed_time=*/seed_times[batch_idx],
+                          /*time=*/dst_time_data,
+                          /*dst_mapper=*/dst_mapper,
+                          /*generator=*/generator,
+                          /*out_global_dst_nodes=*/dst_sampled_nodes);
+                    }
                   }
                 }
               }
@@ -799,7 +900,8 @@ neighbor_sample_kernel(const at::Tensor& rowptr,
                        const at::Tensor& col,
                        const at::Tensor& seed,
                        const std::vector<int64_t>& num_neighbors,
-                       const c10::optional<at::Tensor>& time,
+                       const c10::optional<at::Tensor>& node_time,
+                       const c10::optional<at::Tensor>& edge_time,
                        const c10::optional<at::Tensor>& seed_time,
                        const c10::optional<at::Tensor>& edge_weight,
                        bool csc,
@@ -810,8 +912,8 @@ neighbor_sample_kernel(const at::Tensor& rowptr,
                        bool return_edge_id) {
   const auto out = [&] {
     DISPATCH_SAMPLE(replace, directed, disjoint, return_edge_id, rowptr, col,
-                    seed, num_neighbors, time, seed_time, edge_weight, csc,
-                    temporal_strategy);
+                    seed, num_neighbors, node_time, edge_time, seed_time,
+                    edge_weight, csc, temporal_strategy);
   }();
   return std::make_tuple(std::get<0>(out), std::get<1>(out), std::get<2>(out),
                          std::get<3>(out), std::get<4>(out), std::get<5>(out));
@@ -830,7 +932,8 @@ hetero_neighbor_sample_kernel(
     const c10::Dict<rel_type, at::Tensor>& col_dict,
     const c10::Dict<node_type, at::Tensor>& seed_dict,
     const c10::Dict<rel_type, std::vector<int64_t>>& num_neighbors_dict,
-    const c10::optional<c10::Dict<node_type, at::Tensor>>& time_dict,
+    const c10::optional<c10::Dict<node_type, at::Tensor>>& node_time_dict,
+    const c10::optional<c10::Dict<rel_type, at::Tensor>>& edge_time_dict,
     const c10::optional<c10::Dict<node_type, at::Tensor>>& seed_time_dict,
     const c10::optional<c10::Dict<rel_type, at::Tensor>>& edge_weight_dict,
     bool csc,
@@ -841,8 +944,8 @@ hetero_neighbor_sample_kernel(
     bool return_edge_id) {
   DISPATCH_SAMPLE(replace, directed, disjoint, return_edge_id, node_types,
                   edge_types, rowptr_dict, col_dict, seed_dict,
-                  num_neighbors_dict, time_dict, seed_time_dict,
-                  edge_weight_dict, csc, temporal_strategy);
+                  num_neighbors_dict, node_time_dict, edge_time_dict,
+                  seed_time_dict, edge_weight_dict, csc, temporal_strategy);
 }
 
 std::tuple<at::Tensor, at::Tensor, std::vector<int64_t>>
@@ -850,7 +953,8 @@ dist_neighbor_sample_kernel(const at::Tensor& rowptr,
                             const at::Tensor& col,
                             const at::Tensor& seed,
                             const int64_t num_neighbors,
-                            const c10::optional<at::Tensor>& time,
+                            const c10::optional<at::Tensor>& node_time,
+                            const c10::optional<at::Tensor>& edge_time,
                             const c10::optional<at::Tensor>& seed_time,
                             const c10::optional<at::Tensor>& edge_weight,
                             bool csc,
@@ -860,8 +964,8 @@ dist_neighbor_sample_kernel(const at::Tensor& rowptr,
                             std::string temporal_strategy) {
   const auto out = [&] {
     DISPATCH_DIST_SAMPLE(replace, directed, disjoint, rowptr, col, seed,
-                         {num_neighbors}, time, seed_time, edge_weight, csc,
-                         temporal_strategy);
+                         {num_neighbors}, node_time, edge_time, seed_time,
+                         edge_weight, csc, temporal_strategy);
   }();
   return std::make_tuple(std::get<2>(out), std::get<3>(out).value(),
                          std::get<6>(out));